requirements.txt
beautifulsoup4
pandas
openpyxl
apache-airflow
pydrive
google-auth
google-auth-oauthlib
google-auth-httplib2
selenium
webdriver_manager